+++ .env
@@ -1,0 +1,1 @@
+process.env.REACT_APP_ETH_RPC_URL=http://127.0.0.1:8545
